requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.37.26
botocore==1.37.26
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2
django-oauth-toolkit==3.0.1
django-storages==1.14.6
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
filelock==3.18.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.30.0
whitenoise==6.9.0
